fixed bug with download sync

+++ src/utils/update_calendars.js
@@ -3,14 +3,12 @@ import fs from 'fs';
 import { loadPersist } from '../database/dbManage.js';
 import { addCalendar } from './calendars/calendar.js';
 
-fs.rmSync('calendars/calFiles/', {recursive: true, force: true});
-fs.mkdir('calendars/calFiles', (error) => {
-    if (error) {
-      console.log(error);
-    } else {
-      console.log("New directory created successfully!");
-    }
-});
+if (fs.existsSync('calendars/calFiles/')) {
+  fs.rmSync('calendars/calFiles/', {recursive: true, force: true});
+}
+
+fs.mkdirSync('calendars/calFiles', { recursive: true });
+
 let data = loadPersist();
 
 const timer = ms => new Promise(res => setTimeout(res, ms));
